+++ pybigtools/src/raster.rs
